Some refactoring #64
