+++ MANIFEST.in
@@ -2,5 +2,7 @@ include requirements/*
 include README.md
 include pipestat/schemas/*
 include pipestat/backends/*
+include pipestat/backends/file_backend/*
+include pipestat/backends/db_backend/*
 include pipestat/pipestatreader/*
 include pipestat/jinja_templates/*
