+++ pipestat/backends/db_backend/dbbackend.py
@@ -14,8 +14,8 @@ from sqlmodel import Session, create_engine, select as sql_select
 #     pass
 
 from pipestat.helpers import *
-from pipestat.helpersdb import *
-from .abstract import PipestatBackend
+from pipestat.backends.db_backend.db_helpers import *
+from pipestat.backends.abstract import PipestatBackend
 
 if int(sys.version.split(".")[1]) < 9:
     from typing import List, Dict, Any, Optional, Union
